Remove adapter model helpers from tests
