--- tests/adapter_models/__init__.py
@@ -1,20 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-adapter_models
-
-Test helpers exposing adapter-backed models.
-
-Version:0.1.0
-Author: Timur Kady
-Email: timurkady@yandex.com
-"""
-
-from __future__ import annotations
-
-from .registry import AdapterModelsRegistry, models, registry
-
-__all__ = ["AdapterModelsRegistry", "registry", "models"]
-
-
-# The End
-
